[Example] ggml: simplify llama examples
