+++ wasmedge-ggml/chatml/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "wasmedge-ggml-chatml"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+serde_json = "1.0"
+wasi-nn = { git = "https://github.com/second-state/wasmedge-wasi-nn", branch = "ggml" }
